Added pick up feature
Purpose of this commit:
I added the pick up feature for objects. Objects require two attributes.
One in the tag they need to add the "canPickUp"-Tag in order to be
picked up. The other is the possession of the Target script for
highlighting the object. I added the PickUpScript and modified the
scene by adding additional cameras to look at objects. I had to make
some changes in the project settings since the Input Systems differed.